SQL Code for Database Created
made a basic database with 4 tables, tracking users - this will be linked with an account creation system via php in another iteration. also stores product types and their stock as well as user orders which utilises the other tables to create user order numbers and forms. This is all in 3NF.