user registration / login system working
working account creation system allows users to register an account - these details will then be sent and stored in the database table 'users' (passwords are hashed in the table). User can then log into the existing account and be redirected to the home page. Additional changes have been made to the db.sql file to supplement changes to the php code.